Added transfer queue and change the cmdBuf creation
The command buffer are now created on the fly at every frame.
I have added a command pool for every swapchain frame. This way
we can reset all buffers at the same time once the frame is rendered.
It also reuses the same command buffers instead of creating new ones.
The command pool was also modified to allow reallocations if more
buffers are needed.

I am now using a transfer queue for copy/transition instead of using
the graphic queue. Right now the syncronization is pretty dumb and is
simply waiting for idle device, which tottaly defeat the purposes of
using a separate queue. I will try to modify that in the future...
over